Balance tweaks and code quality improvements
- Increase the nitro engine force bonus for light and medium karts in proportion to their increased consumption (the increased max-speed boost is useful but not always easy to reach, especially if no skidding is used)
- Boost slightly the engine force bonus from nitro
- Slightly tweak the reduce-turn-min factor for heavies as they have a 36% worse turn radius than light, rather than the initial 40%. This keeps 'skidding straight' comparable between the main kart classes.
- Make more startup boost parameters tweakable in kart_characteritics.xml
- Correct a bizarre design choice having startup boost properties checked in multiple code locations
- Tweak the startup boost to make it less kart-class dependent and slightly weaker
- Make the startup boost additional engine force depend on the tier of startup boost, tune it down

+++ src/network/protocols/server_lobby.hpp
@@ -400,7 +400,7 @@ public:
     virtual bool isRacing() const OVERRIDE { return m_state.load() == RACING; }
     bool allowJoinedPlayersWaiting() const;
     void setSaveServerConfig(bool val)          { m_save_server_config = val; }
-    float getStartupBoostOrPenaltyForKart(uint32_t ping, unsigned kart_id);
+    uint8_t getStartupBoostOrPenaltyForKart(uint32_t ping, unsigned kart_id);
     int getDifficulty() const                   { return m_difficulty.load(); }
     int getGameMode() const                      { return m_game_mode.load(); }
     int getLobbyPlayers() const              { return m_lobby_players.load(); }
